pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "torch>=2.2.0", "ninja"]
build-backend = "setuptools.build_meta"

[project]
name = "tide-inference"
version = "0.2.0"
description = "Token-Informed Depth Execution — dynamic per-token layer skipping for transformer inference"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "RightNow AI"},
]
keywords = ["llm", "inference", "early-exit", "transformer", "optimization", "cuda", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: C++",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "datasets",
    "numpy",
    "pyyaml",
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov"]
bench = ["vllm>=0.7.0", "flash-attn>=2.7.0"]

[project.urls]
Homepage = "https://www.rightnowai.co/"
Repository = "https://github.com/RightNow-AI/TIDE"
Issues = "https://github.com/RightNow-AI/TIDE/issues"

[tool.setuptools.package-dir]
TIDE = "python/TIDE"

[tool.setuptools.packages.find]
where = ["python"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
